admin and user routes defined
