+++ components/Head.js
@@ -1,0 +1,15 @@
+import Head from 'next/head'
+import React from 'react'
+
+const _Head = () => {
+  return (
+    <div>
+    <Head>
+      <title>Artrypto</title>
+      <link rel="icon" href="/artrypto-icon.jpg" />
+    </Head>
+  </div>
+  )
+}
+
+export default _Head
