+++ pages/admin/index.js
@@ -1,0 +1,15 @@
+import Head from 'next/head'
+
+export default function Home() {
+    return (
+        <div>
+           
+
+            <main>
+                <h1 className="text-center">
+                    Welcome to Artrypto!
+                </h1>
+            </main>
+        </div>
+    )
+}
