+++ pages/admin/signup.js
@@ -1,26 +1,31 @@
 import React, {useState} from 'react'
 import {Button, Form} from 'react-bootstrap'
-import {useAuth} from "../context/AuthContext";
-import {useRouter} from "next/router";
-import {useSnackbar} from "../context/SnackbarContextProvider";
+import {useAuth} from "../../context/AuthContext";
+import {useSnackbar} from "../../context/SnackbarContextProvider";
 
-const Login = () => {
+const Signup = () => {
 
-    const {login} = useAuth()
-    const router = useRouter()
+    const {signUp} = useAuth()
     const {show} = useSnackbar()
 
     const [data, setData] = useState({
         email: '',
         password: '',
+        username: '',
     })
 
-    const handleLogin = e => {
+    const handleSignup = e => {
         e.preventDefault()
-        login(data.email, data.password)
-            .then(() => {
-                show("Login successful!");
-                router.push('/dashboard')
+
+        if (data.username === '' || data.email === '' || data.password === '') {
+            show("Please fill all the fields.", "error");
+            return;
+        }
+
+        signUp(data.email, data.password, data.username)
+            .then(res => {
+                console.log("Sign up successful", res)
+                show("Sign up successful!");
             })
             .catch(err => console.log(err))
     }
@@ -32,11 +37,29 @@ const Login = () => {
                 margin: 'auto',
             }}
         >
-            <h1 className="text-center my-3 ">Login</h1>
-            <Form onSubmit={handleLogin}>
+            <h1 className="text-center my-3 ">Signup</h1>
+            <Form onSubmit={handleSignup}>
+
+                <Form.Group className="mb-3" controlId="formBasicName">
+                    <Form.Label>Username</Form.Label>
+                    <Form.Control
+                        type="text"
+                        placeholder="Enter username"
+                        onChange={e =>
+                            setData({
+                                ...data,
+                                username: e.target.value,
+                            })
+                        }
+                        value={data.username}
+                    />
+                </Form.Group>
+
                 <Form.Group className="mb-3" controlId="formBasicEmail">
                     <Form.Label>Email address</Form.Label>
                     <Form.Control
+                        type="email"
+                        placeholder="Enter email"
                         onChange={e =>
                             setData({
                                 ...data,
@@ -44,15 +67,14 @@ const Login = () => {
                             })
                         }
                         value={data.email}
-                        required
-                        type="email"
-                        placeholder="Enter email"
                     />
                 </Form.Group>
 
                 <Form.Group className="mb-3" controlId="formBasicPassword">
                     <Form.Label>Password</Form.Label>
                     <Form.Control
+                        type="password"
+                        placeholder="Password"
                         onChange={e =>
                             setData({
                                 ...data,
@@ -60,17 +82,15 @@ const Login = () => {
                             })
                         }
                         value={data.password}
-                        required
-                        type="password"
-                        placeholder="Password"
                     />
                 </Form.Group>
+
                 <Button variant="primary" type="submit">
-                    Login
+                    Signup
                 </Button>
             </Form>
         </div>
     )
 }
 
-export default Login
+export default Signup
