--- pages/dashboard.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Dashboard = () => {
-    return (
-        <div className="text-center mt-5">
-            <h1>This route is protected</h1>
-        </div>
-    )
-}
-
-export default Dashboard
